atomic is protecting us
