+++ app/src/androidTest/java/com/monkeyteam/chimpagne/newtests/model/event/JoinEventTests.kt
@@ -49,7 +49,7 @@ class JoinEventTests {
     assertEquals(ChimpagneRole.NOT_IN_EVENT, event.getRole(anotherAccount.firebaseAuthUID))
 
     var loading = true
-    eventManager.addGuest(
+    eventManager.atomic.addGuest(
         eventId,
         anotherAccount.firebaseAuthUID,
         { loading = false },
